Use float64

+++ slidingwindow_test.go
@@ -34,14 +34,14 @@ func TestAdd(t *testing.T) {
 	sw := &SlidingWindow{
 		window:      2 * time.Second,
 		granularity: time.Second,
-		samples:     []int64{1, 1},
+		samples:     []float64{1, 1},
 		pos:         1,
 		size:        2,
 	}
 
 	sw.Add(1)
 	if v := sw.samples[1]; v != 2 {
-		t.Errorf("expected the 2nd sample to be 2, but got %d", v)
+		t.Errorf("expected the 2nd sample to be 2, but got %f", v)
 	}
 }
 
@@ -49,7 +49,7 @@ func TestAverage(t *testing.T) {
 	sw := &SlidingWindow{
 		window:      10 * time.Second,
 		granularity: time.Second,
-		samples:     []int64{1, 2, 5, 0, 0, 0, 0, 0, 4, 0},
+		samples:     []float64{1, 2, 5, 0, 0, 0, 0, 0, 4, 0},
 		pos:         2,
 		size:        10,
 	}
@@ -79,14 +79,14 @@ func TestReset(t *testing.T) {
 	sw := MustNew(2*time.Second, time.Second)
 	defer sw.Stop()
 
-	sw.samples = []int64{1, 2}
+	sw.samples = []float64{1, 2}
 	sw.pos = 1
 	sw.size = 10
 
 	sw.Reset()
 	for _, v := range sw.samples {
 		if v != 0 {
-			t.Fatalf("expected the samples all to be 0, but at least one value was %d", v)
+			t.Fatalf("expected the samples all to be 0, but at least one value was %f", v)
 		}
 	}
 }
@@ -106,28 +106,28 @@ func TestTotal(t *testing.T) {
 	sw := &SlidingWindow{
 		window:      10 * time.Second,
 		granularity: time.Second,
-		samples:     []int64{1, 2, 5, 0, 0, 0, 0, 0, 4, 0},
+		samples:     []float64{1, 2, 5, 0, 0, 0, 0, 0, 4, 0},
 		pos:         2,
 		size:        10,
 	}
 
 	if v, _ := sw.Total(0); v != 0 {
-		t.Errorf("expected the total with a window of 0 seconds to be 0, not %d", v)
+		t.Errorf("expected the total with a window of 0 seconds to be 0, not %f", v)
 	}
 	if v, _ := sw.Total(time.Second); v != 2 {
-		t.Errorf("expected the total over the last second to be 2, not %d", v)
+		t.Errorf("expected the total over the last second to be 2, not %f", v)
 	}
 	if v, _ := sw.Total(2 * time.Second); v != 3 {
-		t.Errorf("expected the total over the last 2 seconds to be 3, not %d", v)
+		t.Errorf("expected the total over the last 2 seconds to be 3, not %f", v)
 	}
 	if v, _ := sw.Total(4 * time.Second); v != 7 {
-		t.Errorf("expected the total over the last 4 seconds to be 7, not %d", v)
+		t.Errorf("expected the total over the last 4 seconds to be 7, not %f", v)
 	}
 	if v, _ := sw.Total(10 * time.Second); v != 12 {
-		t.Errorf("expected the total over the last 10 seconds to be 12, not %d", v)
+		t.Errorf("expected the total over the last 10 seconds to be 12, not %f", v)
 	}
 	// This one should be equivalent to 10 seconds.
 	if v, _ := sw.Total(20 * time.Second); v != 12 {
-		t.Errorf("expected the total over the last 10 seconds to be 12, not %d", v)
+		t.Errorf("expected the total over the last 10 seconds to be 12, not %f", v)
 	}
 }
